+++ fluxion-stream/src/emit_when.rs
@@ -6,7 +6,7 @@ use crate::types::CombinedState;
 use crate::FluxionStream;
 use fluxion_core::into_stream::IntoStream;
 use fluxion_core::lock_utilities::lock_or_recover;
-use fluxion_core::{StreamItem, Timestamped};
+use fluxion_core::{ComparableSync, StreamItem};
 use fluxion_ordered_merge::OrderedMergeExt;
 use futures::{Stream, StreamExt};
 use std::fmt::Debug;
@@ -17,10 +17,10 @@ use std::sync::{Arc, Mutex};
 ///
 /// This operator gates a source stream based on conditions from a filter stream,
 /// emitting source values only when the combined state passes a predicate.
-pub trait EmitWhenExt<T>: Stream<Item = fluxion_core::StreamItem<T>> + Sized
+pub trait EmitWhenExt<T>: Stream<Item = StreamItem<T>> + Sized
 where
-    T: Timestamped + Clone + Debug + Ord + Send + Sync + Unpin + 'static,
-    T::Inner: Clone + Debug + Ord + Send + Sync + Unpin + 'static,
+    T: ComparableSync,
+    T::Inner: Clone + Debug + Ord + Send + Sync,
 {
     /// Emits source stream values only when the filter condition is satisfied.
     ///
@@ -124,9 +124,9 @@ type IndexedStream<T> =
     Pin<Box<dyn Stream<Item = (fluxion_core::StreamItem<T>, usize)> + Send + Sync>>;
 impl<T, S> EmitWhenExt<T> for S
 where
-    S: Stream<Item = fluxion_core::StreamItem<T>> + Send + Sync + 'static,
-    T: Timestamped + Clone + Debug + Ord + Send + Sync + Unpin + 'static,
-    T::Inner: Clone + Debug + Ord + Send + Sync + Unpin + 'static,
+    S: Stream<Item = StreamItem<T>> + Send + Sync + Unpin + 'static,
+    T: ComparableSync,
+    T::Inner: Clone + Debug + Ord + Send + Sync,
 {
     fn emit_when<IS>(
         self,