+++ fluxion-stream/src/lib.rs
@@ -696,7 +696,5 @@ pub use merge_with::MergedStream;
 pub use ordered_merge::OrderedStreamExt;
 pub use take_latest_when::TakeLatestWhenExt;
 pub use take_while_with::TakeWhileExt;
-pub use types::{
-    CombinedState, OrderedInner, OrderedInnerUnwrapped, OrderedStreamItem, WithPrevious,
-};
+pub use types::{CombinedState, WithPrevious};
 pub use with_latest_from::WithLatestFromExt;